Refactor based on review
