+++ entity/command/executor.go
@@ -1,4 +1,4 @@
-package entity_command
+package command
 
 // Executor interface defines command executor signature
 type Executor interface {
